+++ src/Event/Hook_Event_Config_Interface.php
@@ -3,13 +3,15 @@ declare( strict_types=1 );
 
 namespace Rarst\Laps\Event;
 
+use Rarst\Laps\Record\Collector\Hook_Collector;
+
 /**
  * Interface to retrieve configuration array of hook events to track.
  */
 interface Hook_Event_Config_Interface {
 
 	/**
-	 * @see Hook_Record_Collector::add_event()
+	 * @see Hook_Collector::add_event()
 	 *
 	 * @return array[]
 	 */