+++ src/Record/Collector/Record_Collector_Interface.php
@@ -1,7 +1,9 @@
 <?php
 declare( strict_types=1 );
 
-namespace Rarst\Laps\Record;
+namespace Rarst\Laps\Record\Collector;
+
+use Rarst\Laps\Record\Record_Interface;
 
 /**
  * Interface to retrieve a set of records form collector.