+++ tests/Record/HookTest.php
@@ -4,7 +4,7 @@ declare( strict_types=1 );
 namespace Rarst\Laps\Tests\Record;
 
 use Rarst\Laps\Event\Core_Events;
-use Rarst\Laps\Record\Hook_Record_Collector;
+use Rarst\Laps\Record\Collector\Hook_Collector;
 use Rarst\Laps\Tests\LapsTestCase;
 use Symfony\Component\Stopwatch\Stopwatch;
 
@@ -13,7 +13,7 @@ class HookTest extends LapsTestCase {
 	public function testCollector() {
 
 		$stopwatch = new Stopwatch();
-		$collector = new Hook_Record_Collector( $stopwatch, [ 'core' => new Core_Events() ] );
+		$collector = new Hook_Collector( $stopwatch, [ 'core' => new Core_Events() ] );
 
 		$this->assertTrue( $stopwatch->isStarted( 'Plugins Load' ) );
 		$this->assertTrue( has_action( 'after_setup_theme', [ $collector, 'after_setup_theme' ] ) );