+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 ### Added
-- call backtraces to SQL records
+- call backtraces to SQL records and HTTP calls
 - new format of hook events
 - sidebar tracking